Release 0.4.1 — Twinguard temperature decimals (#352)
AirQualityLevelService.temperature cast the reading with int(), discarding
the decimal Bosch sends (OpenAPI AirQualityLevelServiceStates.temperature =
number). Twinguard temperature reported stepwise integer jumps and appeared
to flat-line. Return the full float; SHCTwinguard.temperature return type
int -> float. Adds service- and model-level regression tests for decimals.

Addresses #352

+++ boschshcpy/models_impl.py
@@ -2516,7 +2516,7 @@ class SHCTwinguard(SHCBatteryDevice):
         return self._airqualitylevel_service.combinedRating
 
     @property
-    def temperature(self) -> int:
+    def temperature(self) -> float:
         return self._airqualitylevel_service.temperature
 
     @property
